move test code to integration tests
